Refactor project initialization and settings
Add browser reload middleware

Update dependencies in pyproject.toml

Add debugger tags to application.html

+++ hooks/post_gen_project.py
@@ -6,40 +6,41 @@ from os import system
 from json import load
 from shutil import which
 from platform import system as platform
+from typing import Literal
 
 # Cookiecutter variables
-stack = "{{ cookiecutter.stack }}"
-css = "{{ cookiecutter.css }}"
+stack: Literal["THAD", "DIRT"] = "{{ cookiecutter.stack }}"
+css: Literal["tailwind", "spruce", "bootstrap"] = "{{ cookiecutter.css }}"
 init_git: bool = {{cookiecutter.init_git}}
 install_packages: bool = {{cookiecutter.install_packages}}
 install_extensions: bool = {{cookiecutter.install_extensions}}
 
 # Create .env file
 Path(".env").write_text(
-    "DEBUG=on\n" 'SECRET_KEY="secret"\n' "DATABASE_URL=sqlite:///dev.sqlite\n"
+    "DEBUG=on\n"
+    'SECRET_KEY="secret"\n'
+    "DATABASE_URL=sqlite:///dev.sqlite\n"
 )
 
 # Handle stack updates
-match stack:
-    case "THAD":
-        Path("resources/js").mkdir()
-    case "DIRT":
-        Path("resources/pages").mkdir()
-        Path("resources/components").mkdir()
-        Path("resources/hooks").mkdir()
-        Path("apps/cpt/management/commands/createview.py").unlink()
-        Path("apps/cpt/templates/cpt/view.html.template").unlink()
+if stack == "THAD":
+    Path("resources/js").mkdir()
+else:
+    Path("resources/pages").mkdir()
+    Path("resources/components").mkdir()
+    Path("resources/hooks").mkdir()
+    Path("apps/cpt/management/commands/createview.py").unlink()
+    Path("apps/cpt/templates/cpt/view.html.template").unlink()
 
 # Handle CSS updates
-match css:
-    case "tailwind":
-        Path("resources/scss/index.scss").unlink()
-        Path("resources/scss").rmdir()
-    case _:
-        Path("tailwind.config.mjs").unlink()
-        Path("postcss.config.cjs").unlink()
-        Path("resources/css/tailwind.css").unlink()
-        Path("resources/css").rmdir()
+if css == "tailwind":
+    Path("resources/scss/index.scss").unlink()
+    Path("resources/scss").rmdir()
+else:
+    Path("tailwind.config.mjs").unlink()
+    Path("postcss.config.cjs").unlink()
+    Path("resources/css/tailwind.css").unlink()
+    Path("resources/css").rmdir()
 
 # Install things if chosen
 if install_packages:
@@ -53,9 +54,14 @@ if install_extensions:
     if which("code") is None:
         print("VSCode not found. Skipping extension installation.")
     else:
-        extensions = load(Path(".vscode/extensions.json").open())["recommendations"]
-        for recommendation in extensions:
-            system(f"code --install-extension {recommendation} --force")
+        try:
+            [
+                system(f"code --install-extension {recommendation} --force")
+                for recommendation in
+                load(Path(".vscode/extensions.json").open())["recommendations"]
+             ]
+        except Exception:
+            print("Error on Reading Recommendations. Please Install Manually.")
 
 if init_git:
     system("git init")
@@ -69,10 +75,7 @@ if init_git:
 
     system("git add -A")
 
-    while (first_commit := input("Create first commit? [y/n] ").lower()) not in [
-        "y",
-        "n",
-    ]:
+    while (first_commit := input("Create first commit? [y/n] ").lower()) not in ("y", "n",):
         print("Please enter 'y' or 'n'")
     if first_commit == "y":
         system('git commit -m ":rocket: Initial commit"')
@@ -81,3 +84,5 @@ else:
     Path(".gitignore").unlink()
     Path(".gitattributes").unlink()
     Path(".pre-commit-config.yaml").unlink()
+
+print("Finished Initializing Project!")
